Gix guis

+++ gui/src/main/java/io/github/stuff_stuffs/tbcexgui/client/api/GuiTextRenderer.java
@@ -3,9 +3,9 @@ package io.github.stuff_stuffs.tbcexgui.client.api;
 import net.minecraft.text.OrderedText;
 
 public interface GuiTextRenderer {
-    double getHeight();
+    int getHeight();
 
-    double getWidth(OrderedText text);
+    int getWidth(OrderedText text);
 
     void render(OrderedText text, int colour, boolean shadow, int backgroundColour);
 }
